+++ packages/botonic-react/src/util/webviews.ts
@@ -1,0 +1,9 @@
+import { params2queryString } from '@botonic/core'
+
+export function generateWebviewUrlWithParams(
+  webview: any,
+  params: any = ''
+): string {
+  const webviewParams = params ? params2queryString(params) : ''
+  return `/webviews/${webview.name}?${webviewParams}`
+}